requirements.txt
transformers==4.26.1
torch==1.13.1
datasets==2.10.1
Pillow==9.4.0
pyarrow
pandas
numpy